# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Kargo",
        "title": "Kargo API",
        "contact": {},
        "version": "v1alpha1"
    },
    "basePath": "/",
    "paths": {
...
        "/v1beta1/projects/{project}/configmaps/{configmap}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a ConfigMap by name from a project's namespace.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Core",
                    "Generic Config",
                    "Project-Level"
                ],
                "summary": "Retrieve a project-level ConfigMap",
                "operationId": "GetProjectConfigMap",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ConfigMap name",
                        "name": "configmap",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ConfigMap resource (k8s.io/api/core/v1.ConfigMap)",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Replace a ConfigMap in a project's namespace. All existing data\nis replaced. Returns the updated Kubernetes ConfigMap resource.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Core",
                    "Generic Config",
                    "Project-Level"
                ],
                "summary": "Replace a project-level ConfigMap",
                "operationId": "UpdateProjectConfigMap",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ConfigMap name",
                        "name": "configmap",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ConfigMap",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateConfigMapRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ConfigMap resource (k8s.io/api/core/v1.ConfigMap)",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a ConfigMap from a project's namespace.",
                "tags": [
                    "Core",
                    "Generic Config",
                    "Project-Level"
                ],
                "summary": "Delete a project-level ConfigMap",
                "operationId": "DeleteProjectConfigMap",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ConfigMap name",
                        "name": "configmap",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Patch a ConfigMap in a project's namespace. Merges provided data\nwith existing data. Use removeKeys to delete specific keys.\nReturns the updated Kubernetes ConfigMap resource.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Core",
                    "Generic Config",
                    "Project-Level"
                ],
                "summary": "Patch a project-level ConfigMap",
                "operationId": "PatchProjectConfigMap",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name",
                        "name": "project",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ConfigMap name",
                        "name": "configmap",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ConfigMap patch",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/PatchConfigMapRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ConfigMap resource (k8s.io/api/core/v1.ConfigMap)",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/v1beta1/shared/configmaps/{configmap}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a shared ConfigMap by name.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Core",
                    "Generic Config",
                    "Shared"
                ],
                "summary": "Retrieve a shared ConfigMap",
                "operationId": "GetSharedConfigMap",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ConfigMap name",
                        "name": "configmap",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ConfigMap resource (k8s.io/api/core/v1.ConfigMap)",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Replace a shared ConfigMap. All existing data is replaced.\nReturns the updated Kubernetes ConfigMap resource.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Core",
                    "Generic Config",
                    "Shared"
                ],
                "summary": "Replace a shared ConfigMap",
                "operationId": "UpdateSharedConfigMap",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ConfigMap name",
                        "name": "configmap",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ConfigMap",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateConfigMapRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ConfigMap resource (k8s.io/api/core/v1.ConfigMap)",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a shared ConfigMap.",
                "tags": [
                    "Core",
                    "Generic Config",
                    "Shared"
                ],
                "summary": "Delete a shared ConfigMap",
                "operationId": "DeleteSharedConfigMap",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ConfigMap name",
                        "name": "configmap",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Patch a shared ConfigMap. Merges provided data\nwith existing data. Use removeKeys to delete specific keys.\nReturns the updated Kubernetes ConfigMap resource.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Core",
                    "Generic Config",
                    "Shared"
                ],
                "summary": "Patch a shared ConfigMap",
                "operationId": "PatchSharedConfigMap",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ConfigMap name",
                        "name": "configmap",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ConfigMap patch",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/PatchConfigMapRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ConfigMap resource (k8s.io/api/core/v1.ConfigMap)",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/v1beta1/system/configmaps/{configmap}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a system-level ConfigMap by name.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Core",
                    "Generic Config",
                    "System-Level"
                ],
                "summary": "Retrieve a system-level ConfigMap",
                "operationId": "GetSystemConfigMap",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ConfigMap name",
                        "name": "configmap",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ConfigMap resource (k8s.io/api/core/v1.ConfigMap)",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Replace a system-level ConfigMap. All existing data is replaced.\nReturns the updated Kubernetes ConfigMap resource.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Core",
                    "Generic Config",
                    "System-Level"
                ],
                "summary": "Replace a system-level ConfigMap",
                "operationId": "UpdateSystemConfigMap",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ConfigMap name",
                        "name": "configmap",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ConfigMap",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateConfigMapRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ConfigMap resource (k8s.io/api/core/v1.ConfigMap)",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a system-level ConfigMap.",
                "tags": [
                    "Core",
                    "Generic Config",
                    "System-Level"
                ],
                "summary": "Delete a system-level ConfigMap",
                "operationId": "DeleteSystemConfigMap",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ConfigMap name",
                        "name": "configmap",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Patch a system-level ConfigMap. Merges provided data\nwith existing data. Use removeKeys to delete specific keys.\nReturns the updated Kubernetes ConfigMap resource.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Core",
                    "Generic Config",
                    "System-Level"
                ],
                "summary": "Patch a system-level ConfigMap",
                "operationId": "PatchSystemConfigMap",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ConfigMap name",
                        "name": "configmap",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ConfigMap patch",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/PatchConfigMapRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ConfigMap resource (k8s.io/api/core/v1.ConfigMap)",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "PatchConfigMapRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "description": {
                    "type": "string"
                },
                "removeKeys": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "UpdateConfigMapRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "description": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token authentication. Obtain token via OIDC/PKCE flow with your identity provider.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}